Introduce log macros
For now, they are simply the aliases to the ESP-IDF log macros. There are many
components that depend on the FreeRTOS, but don't depend on the ESP-IDF itself.
I want to keep these components ESP-IDF agnostic to be able to share them
between different FreeRTOS-based platforms. Maybe, FreeRTOS itself should be
isolated somehow, but for now it's overkill.