Rework task scheduler API
Initially there was an idea, that run() method will be called
periodically to handle asynchronous events. The issue is, that
xEventGroupGetBits() can be used for such scenario, since it doesn't
clear the received bits from the event group. The only way to get and
clear bits is to use xEventGroupWaitBits(), which is used in the wait()
method under the hood.