Rework task scheduling
Initially task scheduling was done via HighResolutionTimer, based on
high-precision ESP timer, which is based on interrupts. This worked
fine, until I integrated DS18B20 sensor. DS18B20 sensor has very
strict timing requirements, which implies that the sensor operations
shouldn't be interrupted by other tasks. DS18B20 is run from the low
priority task and can be interrupted by the high-priority ESP timer.
It leads to the performance degradation as the number of timers growth,
more timers, more alarms are triggered, more likely the DS18B20 sensor
reading will be preempted.

As a result, I've decided to switch to another scheduler, which runs all
tasks in a single FreeRTOS task. There is still a possibility that this
FreeRTOS task can be preempted by another task with a higher priority,
but the probability is much less.